Rename: `time_slices.rs` => `time_slice.rs`
